requirements.txt
pandas==2.2.2
requests==2.32.3
SQLAlchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
